requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.6
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.11.0
django-js-asset==1.2.2
fontawesome-free==5.15.2
pycodestyle==2.6.0
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
